Bug 1383: Tighten version/profile qualification: fail if: requested compat profile && has core profile
On Mesa, if requesting a 3.1 compat profile, we receive a 4.5 core profile.

This is natural due to constraints within glXCreateContextAttribsARB,
i.e. GLX_CONTEXT_PROFILE_MASK_ARB is only a available for versions >= 3.2
and these are not available on Mesa.

Tested with Mesa 18.3.6 of Debian 10 Buster,
which also confirms Bug 1385 fix of limitating GL3CompatNonCompliant
to Mesa < 18.2.0