Graph Font: Add 'name' to Glyph; Drop erroneous Glyph ID_SPACE, ID_CR manual mapping, rely on no-shape for 'space' non-contour.
Drop erroneous Glyph ID_SPACE, ID_CR to rely on no-shape for 'space' non-contour
resolves different cmap-mappings of fonts, not following 'some std'.

Hence getGlyph(glyph_id) no more uses the `font.getGlyph(Glyph.ID_UNKNOWN)` shape,
but a null-shape as intended and using the hmtx and hhea table values for asvance and bounds.

This fixes 'space' spacing in general and specifically FreeSerif-Regular and the like.

This path also simplifies processing/layout of glyphs in process(..) and get*BoundsFU(..).