Bug 805: GraphUI: Add Group.replaceShape(..) and Shape.getParent()
Group.replaceShape(..) allows replacing a shape w/o disturbing a layout, e.g. to zoom one element
by taking it out of a grid-group and placing it on-top of the Scene while using a placeholder in the grid
until returned.

Shape.getParent() - depending on use-case (w/o DAG) - allows access and control of a shape's Group.