Bug 1498: Change Top-Level Widget Mode: Register a top-level Group in Scene, where its zOffset gets adjusted when activated..
.. instead of having a non-working complicated callback orgy setup.

This also takes away the getAdjustedZ() overloading burden (or better uglyness) etc.

Hence Group's setWidgetMode(boolean) became:
- enableTopLevelWidget(Scene)
- disableTopLevelWidget()

The forwardActivation listener is still applied to all children
as well as isActive() is also still overloaded for same required behavior.

However, none of the children is set in 'widget mode'
as well as the Group is simply added to (or removed from)
the Scene's top-level Group list - the holder.

Scene's setActiveShape(Shape) and releaseActiveShape()
handle the top-level Group if affected,
i.e. adding or zero'ing its ZOffset.