Bug 1528: Mitigate OSXUtil.GetNSView() performed on non main-thread
When running with `Main Thread Checker`, enabled by system property
- `nativewindow.debug.OSXUtil.MainThreadChecker`, or
- `nativewindow.debug=all`
OSXUtil.GetNSView() gets notices as it is not forcefully performed on the main-thread.

This change adds an option to GetNSView 'onMainThread'.

On use cases where CreateNSWindow is performed on main-thread anyways,
we utilize CreateNSWindow2, which returns the tuple win and view handle.

Only in one instance we intentionally leave GetNSView performed
on the current thread, MacOSXCGLContext.NSViewDescriptor ctor.
The latter is called from MacOSXCGLContext.NSOpenGLImpl.create
and MacOSXCGLContext.NSOpenGLImpl.makeCurrent and could cause deadlock issues.

Hence this change soothes the debugging experience
and resolves potential future issues in all but one spot.